Внесет правки в моки
