+++ package.json
@@ -28,5 +28,8 @@
   },
   "engines": {
     "node": "18"
+  },
+  "dependencies": {
+    "days": "^1.1.1"
   }
 }
