+++ src/model/comment-model.js
@@ -1,5 +1,6 @@
-import { COMMENTS_LIST_LENGTH } from '../const.js';
-import { createComment } from '../mocks/comment.js';
+import { createComment } from '../mock/comment.js';
+
+const COMMENTS_LIST_LENGTH = 6;
 
 export default class CommentsModel {
   comments = Array.from({ length: COMMENTS_LIST_LENGTH }, createComment);
