+++ src/presenter/main-presenter.js
@@ -1,6 +1,9 @@
 import FilmsPresenter from '../presenter/films-presenter.js';
+import FilmsModel from '../model/film-model.js';
 
 const siteMain = document.querySelector('.main');
-const filmsPresenter = new FilmsPresenter({ filmsContainer: siteMain });
+const filmModel = new FilmsModel();
+
+const filmsPresenter = new FilmsPresenter({ filmsContainer: siteMain, filmModel });
 
 filmsPresenter.init();
