+++ src/util.js
@@ -1,3 +1,7 @@
+import dayjs from 'dayjs';
+
+const DATE_FORMAT = 'D MMMM';
+
 function getRandomInt(max) {
   return Math.floor(Math.random() * max);
 }
@@ -6,4 +10,8 @@ function getRandomArrayElement(items) {
   return items[Math.floor(Math.random() * items.length)];
 }
 
-export { getRandomInt, getRandomArrayElement };
+function humanizeTaskDueDate(dueDate) {
+  return dueDate ? dayjs(dueDate).format(DATE_FORMAT) : '';
+}
+
+export { getRandomInt, getRandomArrayElement, humanizeTaskDueDate };
