+++ src/view/film-card-view.js
@@ -1,19 +1,21 @@
 import { createElement } from '../render.js';
 
-function createFilmCardTemplate() {
+function createFilmCardTemplate(card) {
+  const {title, rating, year, duration, genre, poster, description, comments} = card;
+
   return `
         <article class="film-card">
           <a class="film-card__link">
-            <h3 class="film-card__title">Popeye the Sailor Meets Sindbad the Sailor</h3>
-            <p class="film-card__rating">6.3</p>
+            <h3 class="film-card__title">${title}</h3>
+            <p class="film-card__rating">${rating}</p>
             <p class="film-card__info">
-              <span class="film-card__year">1936</span>
-              <span class="film-card__duration">16m</span>
-              <span class="film-card__genre">Cartoon</span>
+              <span class="film-card__year">${year}</span>
+              <span class="film-card__duration">${duration}</span>
+              <span class="film-card__genre">${genre.join(' ')}</span>
             </p>
-            <img src="./images/posters/popeye-meets-sinbad.png" alt="" class="film-card__poster">
-            <p class="film-card__description">In this short, Sindbad the Sailor (presumably Bluto playing a "role") proclaims himself, in song, to be the greatest sailor, adventurer and…</p>
-            <span class="film-card__comments">0 comments</span>
+            <img src="./images/posters/${poster}" alt="" class="film-card__poster">
+            <p class="film-card__description">${description}</p>
+            <span class="film-card__comments">${comments}</span>
           </a>
           <div class="film-card__controls">
             <button class="film-card__controls-item film-card__controls-item--add-to-watchlist film-card__controls-item--active" type="button">Add to watchlist</button>
@@ -25,8 +27,12 @@ function createFilmCardTemplate() {
 }
 
 export default class FilmCardView {
+  constructor({ film }) {
+    this.film = film;
+  }
+
   getTemplate() {
-    return createFilmCardTemplate();
+    return createFilmCardTemplate(this.film);
   }
 
   getElement() {
